+++ src/geometry/tetra.cpp
@@ -36,7 +36,7 @@ bool Tetra::containsPoint(const std::vector<Point>& vertices, const Point& point
 }
 
 
-bool Tetra::containsEdge(int v1, int v2) const {
+bool Tetra::containsEdge(size_t v1, size_t v2) const {
     
     return (v[0] == v1 && v[1] == v2) || (v[1] == v1 && v[0] == v2) ||
             (v[0] == v1 && v[2] == v2) || (v[2] == v1 && v[0] == v2) ||
